# backend/src/http/handlers/collections/update-collection-customization.ts
import type { FastifyRequest } from "fastify";
import {
    COLLECTION_CUSTOMIZATION_SOURCE_KIND,
    normalizeTraitFilterPresentationConfig,
    type CollectionCustomizationSourceKind,
} from "@artgod/shared/types";
import { ReadModelBadRequestError } from "@artgod/shared/read-models/errors";
import type {
    UpdateCollectionCustomizationInput,
    UpdateCollectionCustomizationOutput,
} from "../../../application/use-cases/collections/update-collection-customization.js";

export type UpdateCollectionCustomizationRoute = {
    Params: {
        chain_ref: string;
        collection_ref: string;
    };
    Body: {
        traitFilterPresentation?: {
            selectedSource?: unknown;
            userConfig?: {
                rangeKeys?: unknown;
            };
        };
    };
};

type MaybePromise<T> = T | Promise<T>;

export class UpdateCollectionCustomizationHttpAdapter {
    constructor(
        readonly updateCollectionCustomizationPort: {
            updateCollectionCustomization(
                input: UpdateCollectionCustomizationInput,
            ): MaybePromise<UpdateCollectionCustomizationOutput>;
        },
    ) {}

    readonly handle = async (
        request: FastifyRequest<UpdateCollectionCustomizationRoute>,
    ) => {
        const input = this.mapRequestToInput(request);
        return await this.updateCollectionCustomizationPort.updateCollectionCustomization(
            input,
        );
    };

    private mapRequestToInput(
        request: FastifyRequest<UpdateCollectionCustomizationRoute>,
    ): UpdateCollectionCustomizationInput {
        const traitFilterPresentation = request.body?.traitFilterPresentation;
        if (!traitFilterPresentation || typeof traitFilterPresentation !== "object") {
            throw new ReadModelBadRequestError(
                "traitFilterPresentation is required",
            );
        }

        const selectedSource = parseSelectedSource(
            traitFilterPresentation.selectedSource,
        );
        const userConfig = normalizeTraitFilterPresentationConfig({
            rangeKeys: parseStringList(
                traitFilterPresentation.userConfig?.rangeKeys,
                "traitFilterPresentation.userConfig.rangeKeys",
            ),
        });

        return {
            chainRef: request.params.chain_ref,
            collectionRef: request.params.collection_ref,
            traitFilterPresentation: {
                selectedSource,
                userConfig,
            },
        };
    }
}

function parseSelectedSource(
    value: unknown,
): CollectionCustomizationSourceKind {
    if (
        value === COLLECTION_CUSTOMIZATION_SOURCE_KIND.User ||
        value === COLLECTION_CUSTOMIZATION_SOURCE_KIND.Extension
    ) {
        return value;
    }
    throw new ReadModelBadRequestError(
        "traitFilterPresentation.selectedSource is invalid",
    );
}

function parseStringList(value: unknown, field: string): string[] {
    if (value === undefined || value === null) {
        return [];
    }
    if (!Array.isArray(value)) {
        throw new ReadModelBadRequestError(`${field} must be an array`);
    }

    return value.map((item) => {
        if (typeof item !== "string") {
            throw new ReadModelBadRequestError(`${field} must contain strings`);
        }
        return item;
    });
}