# docs/backend-api.openapi.yaml
openapi: 3.0.3
info:
    title: ArtGod Backend API
    version: 0.0.1-pre-alpha.1
    description: |
        OpenAPI schema for the current local backend HTTP API.

        Source of truth for routes is `backend/src/http-routes.ts`.
servers:
    - url: http://127.0.0.1:3000
      description: Local desktop backend
...
paths:
...
    /api/{chain_ref}/{collection_ref}/activity:
        get:
            tags: [Activities]
            summary: Get collection activity feed
            operationId: getCollectionActivity
            parameters:
                - $ref: "#/components/parameters/ChainRef"
                - $ref: "#/components/parameters/CollectionRef"
                - $ref: "#/components/parameters/ActivityKindFilter"
                - $ref: "#/components/parameters/MediaMode"
                - $ref: "#/components/parameters/Limit"
                - $ref: "#/components/parameters/Cursor"
                - name: traits
                  in: query
                  required: false
                  description: Repeated `key:value` trait filter (preferred)
                  schema:
                      type: array
                      items:
                          type: string
                  style: form
                  explode: true
                - name: trait
                  in: query
                  required: false
                  description: Repeated `key:value` trait filter (legacy alias)
                  schema:
                      type: array
                      items:
                          type: string
                  style: form
                  explode: true
                - name: trait_ranges
                  in: query
                  required: false
                  description: Repeated `key:from..to` inclusive unsigned-integer range filter
                  schema:
                      type: array
                      items:
                          type: string
                  style: form
                  explode: true
                - name: trait_range
                  in: query
                  required: false
                  description: Repeated `key:from..to` inclusive unsigned-integer range filter (legacy alias)
                  schema:
                      type: array
                      items:
                          type: string
                  style: form
                  explode: true
            responses:
                "200":
                    description: Collection activity feed ordered newest first with bidirectional cursor pagination; `kind=listings` returns collapsed listing rows
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GetCollectionActivityResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/{chain_ref}/{collection_ref}:
        get:
            tags: [Collections]
            summary: Get collection detail with token page and trait facets
            operationId: getCollectionDetail
            parameters:
                - $ref: "#/components/parameters/ChainRef"
                - $ref: "#/components/parameters/CollectionRef"
                - $ref: "#/components/parameters/Limit"
                - $ref: "#/components/parameters/Cursor"
                - $ref: "#/components/parameters/MediaMode"
                - name: token_status
                  in: query
                  required: false
                  description: |
                      Token browser mode.
                      Defaults to `listed`.
                  schema:
                      $ref: "#/components/schemas/TokenBrowserStatus"
                - name: owner
                  in: query
                  required: false
                  description: Optional owner address for holder-scoped token browsing
                  schema:
                      type: string
                - name: traits
                  in: query
                  required: false
                  description: Repeated `key:value` trait filter (preferred)
                  schema:
                      type: array
                      items:
                          type: string
                  style: form
                  explode: true
                - name: trait
                  in: query
                  required: false
                  description: Repeated `key:value` trait filter (legacy alias)
                  schema:
                      type: array
                      items:
                          type: string
                  style: form
                  explode: true
                - name: trait_ranges
                  in: query
                  required: false
                  description: Repeated `key:from..to` inclusive unsigned-integer range filter
                  schema:
                      type: array
                      items:
                          type: string
                  style: form
                  explode: true
                - name: trait_range
                  in: query
                  required: false
                  description: Repeated `key:from..to` inclusive unsigned-integer range filter (legacy alias)
                  schema:
                      type: array
                      items:
                          type: string
                  style: form
                  explode: true
            responses:
                "200":
                    description: Collection detail and paged token results
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GetCollectionDetailResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
                "500":
                    $ref: "#/components/responses/InternalError"
    /api/{chain_ref}/{collection_ref}/customization:
        get:
            tags: [Collections]
            summary: Get collection customization
            operationId: getCollectionCustomization
            parameters:
                - $ref: "#/components/parameters/ChainRef"
                - $ref: "#/components/parameters/CollectionRef"
            responses:
                "200":
                    description: Effective collection customization state
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GetCollectionCustomizationResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
                "500":
                    $ref: "#/components/responses/InternalError"
        put:
            tags: [Collections]
            summary: Update collection customization
            operationId: updateCollectionCustomization
            parameters:
                - $ref: "#/components/parameters/ChainRef"
                - $ref: "#/components/parameters/CollectionRef"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [traitFilterPresentation]
                            additionalProperties: false
                            properties:
                                traitFilterPresentation:
                                    type: object
                                    required: [selectedSource, userConfig]
                                    additionalProperties: false
                                    properties:
                                        selectedSource:
                                            $ref: "#/components/schemas/CollectionCustomizationSource"
                                        userConfig:
                                            $ref: "#/components/schemas/TraitFilterPresentationConfig"
            responses:
                "200":
                    description: Updated effective collection customization state
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GetCollectionCustomizationResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "403":
                    $ref: "#/components/responses/Forbidden"
                "404":
                    $ref: "#/components/responses/NotFound"
                "500":
                    $ref: "#/components/responses/InternalError"
...
components:
    securitySchemes:
        csrfHeader:
            type: apiKey
            in: header
            name: x-artgod-csrf
            description: CSRF header token
        csrfCookie:
            type: apiKey
            in: cookie
            name: artgod_csrf
            description: CSRF cookie token
    parameters:
        ChainRef:
            name: chain_ref
            in: path
            required: true
            schema:
                type: string
            description: Chain reference (slug or public chain id string)
        CollectionRef:
            name: collection_ref
            in: path
            required: true
            schema:
                type: string
            description: Collection slug
...
        Limit:
            name: limit
            in: query
            required: false
            schema:
                type: integer
                minimum: 1
                default: 250
            description: Page size (defaults to 250)
        Cursor:
            name: cursor
            in: query
            required: false
            schema:
                type: string
            description: Opaque page cursor
        ActivityKindFilter:
            name: kind
            in: query
            required: false
            schema:
                $ref: "#/components/schemas/ActivityFeedFilterKind"
            description: |
                Grouped activity-kind filter.
                `sales` returns only `sale`.
                `listings` returns only `listing_created`.
                On the collection activity endpoint, `listings` is collapsed by `tokenId + maker + currency + UTC day` and surfaces the latest raw listing in each group.
                `transfers` returns only `transfer`.
        MediaMode:
            name: media_mode
            in: query
            required: false
            schema:
                type: string
            description: |
                Extension-driven media mode.
                `snapshot` requests canonical token media.
                `artifact` requests extension-provided media when available.
                Unsupported or missing artifact media falls back to canonical token media.
    responses:
        BadRequest:
            description: Invalid request payload or query/path parameter
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        error: bad_request
                        message: Invalid limit
        Forbidden:
            description: Security guard rejected request (host/origin/csrf)
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        error: forbidden
                        message: Missing CSRF header
        NotFound:
            description: Requested entity or route not found
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        error: not_found
                        message: Unknown bootstrap run
...
        InternalError:
            description: Internal server error
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        error: internal_error
                        message: Internal server error
    schemas:
        ErrorResponse:
            type: object
            required: [error, message]
            additionalProperties: false
            properties:
                error:
                    type: string
                message:
                    type: string
        ChainRecord:
            type: object
            required: [id, type, publicChainId, slug, name]
            additionalProperties: false
            properties:
                id:
                    type: integer
                type:
                    type: string
                publicChainId:
                    type: integer
                slug:
                    type: string
                name:
                    type: string
        CollectionStandard:
            type: string
            enum: [erc721, erc1155]
        CollectionStatus:
            type: string
            enum: [bootstrapping, live, paused, disabled]
        CollectionListItem:
            type: object
            required:
                - chainId
                - collectionId
                - slug
                - address
                - standard
                - status
                - deploymentBlock
                - bootstrapAnchorBlock
                - createdAt
                - updatedAt
            additionalProperties: false
            properties:
                chainId:
                    type: integer
                collectionId:
                    type: integer
                slug:
                    type: string
                address:
                    type: string
                standard:
                    $ref: "#/components/schemas/CollectionStandard"
                status:
                    $ref: "#/components/schemas/CollectionStatus"
                deploymentBlock:
                    type: integer
                    nullable: true
                bootstrapAnchorBlock:
                    type: integer
                    nullable: true
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
...
        TokenCursorPage:
            type: object
            required:
                - items
                - nextCursor
                - limit
                - prevCursor
                - totalItems
                - rangeStart
                - rangeEnd
                - currentPage
                - totalPages
            additionalProperties: false
            properties:
                items:
                    type: array
                    items:
                        $ref: "#/components/schemas/TokenCard"
                nextCursor:
                    type: string
                    nullable: true
                limit:
                    type: integer
                prevCursor:
                    type: string
                    nullable: true
                totalItems:
                    type: integer
                rangeStart:
                    type: integer
                rangeEnd:
                    type: integer
                currentPage:
                    type: integer
                totalPages:
                    type: integer
        TokenBrowserStatus:
            type: string
            enum: [listed, all, listed_then_unlisted]
        TraitFilter:
            type: object
            required: [key, value]
            additionalProperties: false
            properties:
                key:
                    type: string
                value:
                    type: string
        TraitRangeFilter:
            type: object
            required: [key, fromValue, toValue]
            additionalProperties: false
            properties:
                key:
                    type: string
                fromValue:
                    type: string
                    nullable: true
                toValue:
                    type: string
                    nullable: true
        TraitFilterDisplayKind:
            type: string
            enum: [set, range]
        TraitFacetValue:
            type: object
            required: [value, tokenCount]
            additionalProperties: false
            properties:
                value:
                    type: string
                tokenCount:
                    type: integer
        TraitFacet:
            type: object
            required: [key, displayKind, minValue, maxValue, values]
            additionalProperties: false
            properties:
                key:
                    type: string
                displayKind:
                    $ref: "#/components/schemas/TraitFilterDisplayKind"
                minValue:
                    type: string
                    nullable: true
                maxValue:
                    type: string
                    nullable: true
                values:
                    type: array
                    items:
                        $ref: "#/components/schemas/TraitFacetValue"
        CollectionCustomizationSource:
            type: string
            enum: [user, extension]
        TraitFilterPresentationConfig:
            type: object
            required: [rangeKeys]
            additionalProperties: false
            properties:
                rangeKeys:
                    type: array
                    items:
                        type: string
...
        CollectionCustomization:
            type: object
            required: [traitFilterPresentation]
            additionalProperties: false
            properties:
                traitFilterPresentation:
                    $ref: "#/components/schemas/TraitFilterPresentationFeatureState"
...
        CollectionMediaMode:
            type: object
            required: [key, label]
            additionalProperties: false
            properties:
                key:
                    type: string
                label:
                    type: string
        CollectionMediaState:
            type: object
            required: [selectedMode, defaultMode, availableModes]
            additionalProperties: false
            properties:
                selectedMode:
                    type: string
                defaultMode:
                    type: string
                availableModes:
                    type: array
                    items:
                        $ref: "#/components/schemas/CollectionMediaMode"
        GetCollectionDetailResponse:
            type: object
            required: [chain, collection, media, traits, tokens]
            additionalProperties: false
            properties:
                chain:
                    $ref: "#/components/schemas/ChainRecord"
                collection:
                    $ref: "#/components/schemas/CollectionListItem"
                media:
                    $ref: "#/components/schemas/CollectionMediaState"
                traits:
                    type: object
                    required: [selected, selectedRanges, facets]
                    additionalProperties: false
                    properties:
                        selected:
                            type: array
                            items:
                                $ref: "#/components/schemas/TraitFilter"
                        selectedRanges:
                            type: array
                            items:
                                $ref: "#/components/schemas/TraitRangeFilter"
                        facets:
                            type: array
                            items:
                                $ref: "#/components/schemas/TraitFacet"
                tokens:
                    $ref: "#/components/schemas/TokenCursorPage"
        GetCollectionCustomizationResponse:
            type: object
            required: [chain, collection, customization]
            additionalProperties: false
            properties:
                chain:
                    $ref: "#/components/schemas/ChainRecord"
                collection:
                    $ref: "#/components/schemas/CollectionListItem"
                customization:
                    $ref: "#/components/schemas/CollectionCustomization"
...
        ActivityScopeKind:
            type: string
            enum: [token, collection, attribute]
        ActivityKind:
            type: string
            enum:
                [
                    transfer,
                    sale,
                    listing_created,
                    listing_cancelled,
                    bid_created,
                    bid_cancelled,
                    custom,
                ]
        ActivityFeedFilterKind:
            type: string
            enum: [sales, listings, transfers]
        ActivitySourceKind:
            type: string
            enum: [onchain, offchain, extension]
...
        ActivityFeedItem:
            type: object
            required:
                [
                    id,
                    scopeKind,
                    kind,
                    contract,
                    tokenId,
                    occurredAt,
                    sourceKind,
                    sourceName,
                    orderId,
                    blockNumber,
                    txHash,
                    logIndex,
                    from,
                    to,
                    maker,
                    taker,
                    side,
                    amount,
                    price,
                    currency,
                    payload,
                    isCollapsed,
                    collapsedEventCount,
                    collapsedWindowStartUtc,
                    collapsedWindowEndUtc,
                ]
            additionalProperties: false
            properties:
                id:
                    type: integer
                scopeKind:
                    $ref: "#/components/schemas/ActivityScopeKind"
                kind:
                    $ref: "#/components/schemas/ActivityKind"
                contract:
                    type: string
                tokenId:
                    type: string
                    nullable: true
                occurredAt:
                    type: integer
                    description: Unix timestamp in seconds
                sourceKind:
                    $ref: "#/components/schemas/ActivitySourceKind"
                sourceName:
                    type: string
                orderId:
                    type: string
                    nullable: true
                blockNumber:
                    type: integer
                    nullable: true
                txHash:
                    type: string
                    nullable: true
                logIndex:
                    type: integer
                    nullable: true
                from:
                    type: string
                    nullable: true
                to:
                    type: string
                    nullable: true
                maker:
                    type: string
                    nullable: true
                taker:
                    type: string
                    nullable: true
                side:
                    type: string
                    enum: [buy, sell]
                    nullable: true
                amount:
                    type: string
                    nullable: true
                price:
                    type: string
                    nullable: true
                currency:
                    type: string
                    nullable: true
                payload:
                    type: object
                    nullable: true
                    additionalProperties: true
                isCollapsed:
                    type: boolean
                    description: True when the item represents a collapsed collection-listings row rather than a single raw activity row
                collapsedEventCount:
                    type: integer
                    nullable: true
                    description: Number of raw listing events represented by a collapsed listings row
                collapsedWindowStartUtc:
                    type: integer
                    nullable: true
                    description: Inclusive UTC-day window start in Unix seconds for a collapsed listings row
                collapsedWindowEndUtc:
                    type: integer
                    nullable: true
                    description: Inclusive UTC-day window end in Unix seconds for a collapsed listings row
        ActivityCursorPage:
            type: object
            required:
                - items
                - nextCursor
                - limit
                - prevCursor
                - totalItems
                - rangeStart
                - rangeEnd
                - currentPage
                - totalPages
            additionalProperties: false
            properties:
                items:
                    type: array
                    items:
                        $ref: "#/components/schemas/ActivityFeedItem"
                nextCursor:
                    type: string
                    nullable: true
                limit:
                    type: integer
                prevCursor:
                    type: string
                    nullable: true
                totalItems:
                    type: integer
                rangeStart:
                    type: integer
                rangeEnd:
                    type: integer
                currentPage:
                    type: integer
                totalPages:
                    type: integer
        ActivityFeedIncludes:
            type: object
            required: [tokensById]
            additionalProperties: false
            properties:
                tokensById:
                    type: object
                    additionalProperties:
                        $ref: "#/components/schemas/TokenPresentationSummary"
        GetCollectionActivityResponse:
            type: object
            required: [chain, collection, media, traits, activities, included]
            additionalProperties: false
            properties:
                chain:
                    $ref: "#/components/schemas/ChainRecord"
                collection:
                    $ref: "#/components/schemas/CollectionListItem"
                media:
                    $ref: "#/components/schemas/CollectionMediaState"
                traits:
                    type: object
                    required: [selected, selectedRanges, facets]
                    additionalProperties: false
                    properties:
                        selected:
                            type: array
                            items:
                                $ref: "#/components/schemas/TraitFilter"
                        selectedRanges:
                            type: array
                            items:
                                $ref: "#/components/schemas/TraitRangeFilter"
                        facets:
                            type: array
                            items:
                                $ref: "#/components/schemas/TraitFacet"
                activities:
                    $ref: "#/components/schemas/ActivityCursorPage"
                included:
                    $ref: "#/components/schemas/ActivityFeedIncludes"
...